+++ packages/teams_memory/CHANGELOG.md
@@ -3,7 +3,7 @@
 ## Latest Changes
 
 ### Features
-- Added method `BaseMemoryModule.get_memories_with_citations` to easily get the messages associated to memories. 
+- Added method `BaseMemoryModule.get_memories_with_attributions` to easily get the messages associated to memories. 
 
 ### Bug Fixes
 - Tech assistant sample changes: fixed a bug, refactored citations, updated it to use python's logger library.